+++ inc/options.php
@@ -1,0 +1,16 @@
+<?php
+
+// if( function_exists('acf_add_options_page') ) {
+//
+//   // Options pour les Actualités
+//   acf_add_options_page(array(
+// 		'page_title' 	=> __( 'Réglages des actualités', 'telabotanica' ),
+// 		'menu_title'	=> __( 'Réglages', 'telabotanica' ),
+// 		'menu_slug' 	=> 'news-settings',
+//     'capability'	=> 'edit_posts',
+//     'parent_slug'	=> 'edit.php',
+//     'icon_url'    => 'dashicons-admin-settings',
+// 		'redirect'		=> false
+// 	));
+//
+// }